+++ .vscode/settings.json
@@ -7,5 +7,8 @@
     "amplify/#current-cloud-backend": true,
     "amplify/backend/amplify-meta.json": true,
     "amplify/backend/awscloudformation": true
-  }
+  },
+  "eslint.workingDirectories": [
+    "./frontend"
+  ]
 }
